[JAVA-46871] Created new module core-java-lang-oop-methods-2 and moved code from core-java-lang-oop-methods
