+++ core-java-modules/core-java-lang-oop-methods-2/src/main/java/com/baeldung/methodoverloadingoverriding/util/Multiplier.java
