+++ core-java-modules/core-java-lang-oop-methods-2/src/main/java/com/baeldung/methods/Vehicle.java
