+++ core-java-modules/core-java-lang-oop-methods-2/src/test/java/com/baeldung/methodoverloadingoverriding/MethodOverridingUnitTest.java
@@ -4,7 +4,8 @@ import com.baeldung.methodoverloadingoverriding.model.Car;
 import com.baeldung.methodoverloadingoverriding.model.Vehicle;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import static org.assertj.core.api.Assertions.*;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class MethodOverridingUnitTest {
