+++ core-java-modules/core-java-lang-oop-methods-2/src/test/java/com/baeldung/methods/VehicleProcessorUnitTest.java
@@ -1,8 +1,8 @@
 package com.baeldung.methods;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
 import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 class VehicleProcessorUnitTest {
