+++ core-java-modules/core-java-lang-oop-methods-2/src/test/java/com/baeldung/testhashcode/MyClass.java
@@ -1,0 +1,15 @@
+package com.baeldung.testhashcode;
+
+public class MyClass {
+    private String value;
+
+    public MyClass(String value) {
+        this.value = value;
+    }
+
+    @Override
+    public int hashCode() {
+        return value != null ? value.hashCode() : 0;
+    }
+
+}
