--- core-java-modules/core-java-lang-oop-methods/src/test/java/com/baeldung/testhashcode/MyClass.java
@@ -1,15 +1,0 @@
-package com.baeldung.testhashcode;
-
-public class MyClass {
-    private String value;
-
-    public MyClass(String value) {
-        this.value = value;
-    }
-
-    @Override
-    public int hashCode() {
-        return value != null ? value.hashCode() : 0;
-    }
-
-}
