+++ core-java-modules/pom.xml
@@ -184,6 +184,7 @@
         <module>core-java-lang-oop-inheritance</module>
         <module>core-java-lang-oop-inheritance-2</module>
         <module>core-java-lang-oop-methods</module>
+        <module>core-java-lang-oop-methods-2</module>
         <module>core-java-lang-oop-others</module>
         <module>core-java-lang-operators</module>
         <module>core-java-lang-operators-2</module>
